Nest `nodes` under `cluster` command in the CLI and prefix node references with cluster in the API.
