+++ cmd/cluster_nodes_unregister.go
@@ -15,18 +15,26 @@
 package cmd
 
 import (
+	"strconv"
+
 	"github.com/spf13/cobra"
 )
 
-func newNodesCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "nodes",
-		Short: "Manage nodes",
+func newClusterNodesUnregisterCmd() *cobra.Command {
+	cmd := cobra.Command{
+		Use:   "unregister [flags] <nodeID>",
+		Short: "Unregister (remove) a node from the cluster [Root Only]",
+		Args:  cobra.ExactArgs(1),
+		Run: func(cmd *cobra.Command, args []string) {
+			node, parseErr := strconv.ParseInt(args[0], 10, 64)
+			exitOnError(cmd, parseErr, "Not valid node id: %s")
+
+			client := NewApiClient(cmd)
+
+			apiError := client.ClusterNodes().Unregister(node, false)
+			exitOnError(cmd, apiError, "Error removing parser: %s")
+		},
 	}
 
-	cmd.AddCommand(newNodesListCmd())
-	cmd.AddCommand(newNodesShowCmd())
-	cmd.AddCommand(newNodesUnregisterCmd())
-
-	return cmd
+	return &cmd
 }
